+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `RawEvent.toCommon` alias function for `Event.toCommon`.
 
 ## 6.1.0 - 2019-05-22
 - Add `CommonEvent` type to allow common operations above all events without any specific data.